kernel: automate process of uploading kernel modules to remote host
Add functionality for uploading the compiles kernel modules to a
remote host.  This allow you to develop locally and push the
modules remotely.  This is especially useful if your module
causes a kernel crash, which you want to avoid on the system you
are running your editor/devel-env.

Q: Want to compile locally and push the binary modules to a remote host.
A: Yes, this is supported.

The Makefile target "push_remote" uploads the kernel module to a
remote host.  (You need to setup SSH-keys to SSH allow root logins.)

Usage example:
 make push_remote kbuilddir=~/git/kernel/net-next/ HOST=192.168.122.49

If you want to run this manually call the script directly:
 ./scripts/push_remote.sh 192.168.122.49

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>