alf_queue_test: testing enqueue until full and dequeue until empty
This test also demonstrates effect of increasing bulk enqueue.

As current enqueue semantics is to abort if the entire bulk does
not fit.  While the bulk dequeue will return the number of
elements it was able to bulk dequeue.

Thus, we might not use the entire capacity of the queue, when
using too large bulk enqueues.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>