mm: qmempool: a quick queue based mempool
A quick or queue-based memory pool, that functions as a cache
in-front of kmem_cache SLAB/SLUB allocators.

A quick queue-based memory pool, that functions as a cache in-front
of kmem_cache SLAB/SLUB allocators.  Which allows faster than
SLAB/SLUB reuse/caching of fixed size memory elements

The speed gain comes from, the shared storage, using a Lock-Free
queue that supports bulking elements with a single cmpxchg.

The Lock-Free queue is based on an array (of pointer to elements).
This make access more cache optimal, as e.g. on 64bit 8 pointers
can be stored per cache-line (which is superior to a linked list
approach).

Only storing the pointers to elements, is also beneficial as we
don't touch the elements data.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>