alf_queue: adding a smp_wmb() in alf_mc_dequeue()
On architechtures with weak Memory Ordering, a memory barrier is
needed in dequeue(), to make sure the dequeue LOADs from the
array happens BEFORE the STORE to q->consumer.tail.

The dequeue LOADs have a dependent STORE into ptr, thus a smp_wmb()
is enough. The paired memory barrier happens in enqueue with the
implicit full-MB in cmpxchg.

We have not been able to provoke this error on any of the tested
architechtures.  Note, we have been able to provoke an error if
the smp_wmb() in enqueue is missing.

Notice for Intel x86 this is only a compiler barrier().

Based on discussions with Hannes.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>
Signed-off-by: Hannes Frederic Sowa <hannes@stressinduktion.org>