alf_queue_concurrency: add benchmark measurement of dequeue cost
Measure the cost of dequeuing an element in the case of
concurrent producers enqueueing.  This should measure the worst
effect producers/enqueuers can cause for a single consumer.

Notice the concurrent producers will be reading consumer.tail,
which is bad for the consumer/dequeuer as it need to write to
this cache line.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>