alf_queue: correct memcpy versions of enqueue/dequeue
This code is based on Hannes work, thus he should be the author
of this commit.

This fixes the memcpy version of the enqueue/dequeue helpers:
 __helper_alf_enqueue_store_memcpy()
 __helper_alf_dequeue_load_memcpy()

But don't enable them by default, because it causes a performance
regression.

When enabled results from alf_queue_bench on a CPU E5-2695:

 Type:ALF-simple Per elem: 46 cycles(tsc) 18.511 ns (step:0)
  - (measurement period time:0.370235788 sec time_interval:370235788)
  - (invoke count:20000000 tsc_interval:925588609)

 Type:ALF-multi Per elem: 46 cycles(tsc) 18.400 ns (step:128)
  - (measurement period time:0.471040922 sec time_interval:471040922)
  - (invoke count:25600000 tsc_interval:1177601335)

These 18.511 ns / 18.400 ns is a regression compared to the
current performance cost of 10.910 ns / 13.576 ns.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>
Signed-off-by: Hannes Frederic Sowa <hannes@stressinduktion.org>