alf_queue: loop unrolled enqueue/dequeue helpers
Implemented:
 __helper_alf_enqueue_store_unroll()
 __helper_alf_dequeue_load_unroll()

Which implements loop unrolling in case of bulking.

Comparing enqueue/dequeue helpers "mask" -> "unroll":

It is interesting to see that none-bulking benchmarks is not hurt,
it is even slightly improved:

 (none-bulking)
 simple: 10.912 ns -> 10.149 ns
 multi : 13.589 ns -> 13.579 ns

The loop unrolling does improve performance, but not much, and
mostly for larger and 4 aligned cases.

 (bulking)
 bulk2 :  5.193 ns -> 5.489 ns
 bulk4 :  3.371 ns -> 3.225 ns
 bulk6 :  2.269 ns -> 2.211 ns
 bulk8 :  1.997 ns -> 1.907 ns
 bulk16:  1.493 ns -> 1.197 ns

I was hoping to see a larger effect of this loop unrolling, not
because of the unroll, but because I was hoping to see a memory
pipeline effect.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>