alf_queue: reduce "AND" operations in enqueue/dequeue helper "mask_less"
It is possible to avoid the masking of the index with q->mask,
if we know that wrapping will not occur.

This is implemented in:
 __helper_alf_enqueue_store_mask_less()
 __helper_alf_dequeue_load_mask_less()

Comparing enqueue/dequeue helpers "mask" -> "mask_less":

It is somewhat surprising to see the improvement for the
none-bulking case.

 (none-bulking)
 simple: 10.912 ns ->  9.438 ns
 multi : 13.589 ns -> 13.411 ns

There is an regression in case of bulking only 2 elements, but
else its a slight improvement.

 (bulking)
 bulk2 :  5.193 ns -> 6.785 ns
 bulk4 :  3.371 ns -> 3.214 ns
 bulk6 :  2.269 ns -> 2.388 ns
 bulk8 :  1.997 ns -> 1.998 ns
 bulk16:  1.493 ns -> 1.399 ns

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>