alf_queue: SPSC: Single-Producer/Consumer variants of enqueue/dequeue
This implements the Single-Producer/Consumer variants the
Enqueue/Dequeue functions.  As the name indicate, only a single CPU is
allowed to use the queue for enqueue, and another single CPU is
allowed to dequeue.

Some assertion statements are added to catch race conditions, that can
only occur if the queue used incorrectly, either by being accessed by
several CPU or due to preemption.  The assertions can be disabled
compile time as they do add a slight overhead.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>