alf_queue: changing return semantics of enqueue
Notice the ALF queue return semantics is different from the DPDK
ring_queue semantics.  Also after this change.

The ALF enqueue function's return value is inconsistent with the
dequeue return value (which returns 0 in case it cannot dequeue).

Change enqueue to return 0 instead of -ENOBUFS, when it cannot enqueue
the elements requested.

Even-though current API have the "fixed" semantics of aborting if it
cannot enqueue the full bulk size.  Users of this API should check on
the returned number of enqueue elements match, to verify enqueue was
successful.  This allow us to introduce a "variable" enqueue scheme
later.

Notice this change also fix up the wrong checks on the return value of
dequeue, which tested on return < 0, which dequeue will never return.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>