alf_queue_helper: yet another enqueue/dequeue helper function "nomask"
Still searching for the optimal helper function, both fast and small in
code size.  Was hoping this would produce smaller code, but it did not.

It is faster for the fast-reuse case, but slower for bulking.
 qmempool:  "unroll"   -> "nomask"
 fastpath  :  9.659 ns ->  9.333 ns
 N-pattern : 13.303 ns -> 14.479 ns

This helper is larger than "mask_less2" (110 bytes):
 helper_alf_enqueue_store_nomask: 127 bytes
 helper_alf_dequeue_load_nomask : 122 bytes

Thus, not enabling this helper by default, still keeping "unroll".

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>