build: handle if alf_queue or qmempool is compiled into kernel
When porting/proposing this upstream, the alf_queue and qmempool will
be compiled into the mainline kernel.  This will cause conflicts with
exported symbols from the local modules.

Detect and handle this situation by matching on the kernel config
options that I'll be proposing upstream.  And then only compile
local modules if they are NOT defined.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>