time_bench: GCC notes on rdtsc and rdtscp
Adding these notes to the code, so we doni't forget.

Hannes found out that __builtin_ia32_rdtsc and
 __builtin_ia32_rdtscp are undocumented available in gcc, so there
is no need to write inline assembler functions for them any more.

 unsigned long long __builtin_ia32_rdtscp(unsigned int *foo);
   (where foo is set to: numa_node << 12 | cpu)
    and
 unsigned long long __builtin_ia32_rdtsc(void);

In the time_bench code we combine the calls with CPUID, thus
I don't see how this is directly appreciable.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>