alf_queue: improve/correct memory barriers for e.g. powerpc
Changing the memory barrier in alf_mc_dequeue(), based on
feedback from Paul E. McKenney, and discussions with Alexander
Duyck and Hannes Frederic Sowa.

And after reading blogpost[1]: "Acquire and Release Semantics"
 [1] http://preshing.com/20120913/acquire-and-release-semantics/

The barrier protection needed in alf_mc_dequeue() relates to
LOADing the elements (from the ring/array) MUST happen before the
STORE to q->consumer.tail, which marks that the loads/dequeue are
done and available for a producer to override.

According to the blogpost[1] terminology this is a #LoadStore
type memory barrier.  And this can be solved with load_acquire
and store_release.

In alf_mc_dequeue() use smp_store_release() when STORE'ing
q->consumer.tail.  The corresponding load_acquire, when reading
consumer.tail, is full-filled in alf_mp_enqueue() by the implicit
full memory barrier of cmpxchg.

For x86 this still translates into compiler barriers, but for
architectures like PowerPC, this improves the situation from
a "sync" to a "__lwsync".

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>
Reviewed-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>