build: extend build system to use .config include
It can be useful to disable/enable certain modules.  Add this
feature by using a simple .config file that gets included.

Specifically; this feature is needed for the module slab_bulk_test01
which is testing the SLAB bulking API, which is not merged into
the mainline kernel.

When .config does not exists then it will be created with
the contents from config.default.

If the future, I have a dream of integrating something
like Buildroot for selection, but for now manual edit/adjust
of .config is needed.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>