slab_bulk_test03: construct worse-case for bulk free
This module provoke the worse-case behavior for kmem_cache_free_bulk(),
which is when adjacent objects in the array belongs to difference pages.

This is worse-case for free_bulk, because it cannot exploit the
oppotunity to coalesce object belonging to the same page.

Notice performance tests also count/measure the overhead of calling
the hash funtion jhash * 2.  Because the algorithm stores the objects
into a hash table and extract objects belonging to a different page
from the last extracted object.

Also added some tests trying to crash the system by misusing the API,
e.g releasing NULL pointers.  And added some tests exercising
different parts of building the detached freelist.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>