slab_bulk_test03: don't create a new kmem_cache inside benchmark call
It is wrong to create and delete a kmem_cache each time the benchmark
function call is invoked, because the test is maintaining a seperate
hash with objects (in-order to create worst-case test).

That meant that function was allocating from one slab and possibly
returning objects from the preallocated slab.  This only worked due to
to SLUB aliasing same size objects.

Fix by only using a single "global" SLAB/kmem_cache.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>