slab_test02: add a new basic slab/kmem_cache test
Basic slab test of create and destroy of a kmem_cache.

This test is very basic, but I discovered a bug in kmem cgroups.

That is, if a kmem_cache is created and destroy'ed too quickly
something in the kmem cgroups system breaks. Specifically in
function memcg_create_kmem_cache().  This seems to be related
to starting/running workqueues for creating a memcg kmem_cache.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>