page_bench03: get page specfic stats back
A trick used by many drivers is to allocate a larger order page,
and partition this larger page into smaller pieces, sometimes
called page fragements.

Here the calculation show that the CPU cycles cost is per 4K
page-size blocks.  As this normalize the result to 4K page-size, for
easier comparison.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>