time_bench: extend concurrency API with passing along opaque "data" pointer
Existing bench functions allow to pass along a void *data pointer,
which is used for opaque handing over some data structure.

For parallel tasks this is a useful feature, as it easier allow us
to pass along a data structure that can be shared among several CPUs
when e.g. benchmarking and testing concurrent access primitives on
this shared data structure. This avoid depending on global variables.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>