skb_array_bench01: simple bench to measure min achievable overhead
This is a really fake benchmark, but it sort of shows the minimum
overhead achievable with this type of queue, where it is the same
CPU enqueuing and dequeuing, and cache is guaranteed to be hot.

Measured on a i7-4790K CPU @ 4.00GHz, the average cost of
enqueue+dequeue of a single object is around 102 cycles(tsc).

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>