skb_array_parallel01: parallel benchmark module for linux/skb_array.h
This benchmark try to keep two (or more) CPUs busy enqueuing or
dequeuing on the same skb_array queue.  It prefills the queue,
and stops the test as soon as queue is empty or full, or
completes a number of "loops"/cycles.

The objects used in the queue are "fake" objects.  Thus, we can
keep an eternal enq+deq cycles without needing to handle objects.
This also keeps the benchmark focus in the queue itself.

The skb_array queue should behave well the case of two CPUs
single producer(enq) and single consumer(deq), as the
producer/consumer part of struct skb_array is kept on seperate
cachelines.  Plus, as soon as the queue size is above 8 (on
64-bit) the array memory usage (8*8) 64 bytes is above one
cache-line, thus the producer/consumer no longer bounce the array
cacheline.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>