alf_queue_parallel01: benchmark for alf_queue similar to skb_array_parallel01
This benchmark is for comparing alf_queue with skb_array in a
concurrency/multi-CPU workload.  Modeled after the
skb_array_parallel01 module.

The alf_queue do have the problem MST points out.  The issue is that
alf_queue "reads" the opposite cacheline of the consumer.tail/producer.tail
to determine if space-is-left/queue-is-empty.  This cause an expensive
transition for the cache coherency protocol.

The test (currently) does not utilize the bulking feature of alf_queue,
which is suppose to mitigate this behavior of bouncing the cacheline.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>