skb_array: compile time warning if skb_array gets upstream
It confused me a bit when I applied the skb_array patches to
my devel kernel tree, as this disabled any change to the
skb_array.h in the prototype-kernel git tree.

To make people/users aware of this, generate a compile time warning
to make it move obvious.  This also help when skb_array and ptr_ring
is accepted in upstream kernels.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>