kernel: sample/bpf eBPF/LLVM compiling via pipe does not catch clang errors
It is problematic that clang compile errors does not get caught by
the Makefile system.  This is because the CLANG emit-llvm output
is piped directly to LLC, and thus make cannot see the exit code of CLANG.
This result in a empty eBPF program, which is fairly confusing.

This patch removes the pipe and generate the eBPF code in two steps.

Notice, choose use clang -S to generate the non-binary LLVM assembly
language format suffix .ll, instead of the LLVM bitcode format (.bc),
in hopes this will allow developers a little more insight into the
transformation steps for debugging issues.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>