samples/bpf: xdp_ddos01 replace option --ifindex with --dev
It is really counter intuitive that the users have to specify an
ifindex number for the network device.  There exists standard
API interfaces like if_nametoindex(3) that can do such conversions.

This patch removes the option --ifindex and replace it with --dev
which allows users to specify the device name instead.  Lets hide
the details that XDP uses the ifindex.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>