samples/bpf: xdp_ddos01 open and close blacklist file to catch XDP reload
When reloading the XDP program, the file handle to the exported/pinned
eBPF map file may belong on the old map and XDP program.  This is something
that we plan to fix, so a XDP reload takes over the old map. But given this
is not implemented, open and close the file between each stats poll, to catch
if the file changed.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>