samples/bpf: xdp_ddos01 use separate exit code for map file missing
Introduce exit code EXIT_FAIL_MAP_FILE which is used then opening
the bpf map file fails (in open_bpf_map()).

This is useful to keep a separate exit code for this, as this can occur
if the XDP program is reloaded, and a shell script might like to detect this
as a separate error code.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>