samples/bpf: xdp_ddos01 add support for port blacklisting
This adds support for keeping an array-based blacklist of ports.
Perf report indicates that when bpf JITing is enabled in the kernel
there is a 3-4% decrease in cpu samples (measured with perf) to using
this over the new inlined bpf hash maps for storing port blacklist data.
Perf report indicates performance is roughly equal comparing this method
to bpf mash maps when bpf JITing is disabled in the kernel.

Signed-off-by: Andy Gospodarek <gospo@broadcom.com>