samples/bpf: maintain a copy of uapi/linux/bpf.h
Getting the enum's for eBPF helpers in bpf_helpers.h creates a
dependency to uapi/linux/bpf.h which define enum bpf_func_id.

It is unfortunately not possible to use #ifdef BPF_FUNC_xxx
because enum's does not get expanded during C preprocessor stage,
see why in [1]:

[1] https://gcc.gnu.org/onlinedocs/gcc-3.0.2/cpp_4.html#SEC38

Thus, the only option "here-and-now" is to maintain a local copy of
the include file uapi/linux/bpf.h in this git repository.

The more long-term solution could be to change the kernels UAPI
file to use defines for every enum, like include/uapi/linux/in.h.
But that will likely never happen ;-)

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>