samples/bpf: attempt getting bpf compiling on Fedora 25
This is an attempt of getting the kernel/samples/bpf part of the
prototype-kernel project compiling on base Fedora 25 distro kernel.

The XDP features needed should be part of kernel 4.9, it is only compile
dependencies to features available in kernel 4.10 that is blocking the
compile of the eBPF program.  Features we don't use.

This commit change:

1. Only warn if make headers_install is missing, it could be that
   distro or installed kernel headers are sufficient.

2. Comment of most of perf-sys.h as most is not needed for eBPF use-case.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>