xdp_bench02: playing with simulation drop patterns
I think I found a way to simulate part of RX-stages, particularly the
effect of handling XDP packets before netstack packets.  Without any
code changes.

The basic idea is to have an XDP program that simply drop half of all
incoming packets. The trick is now to change the drop pattern.
Instead of dropping every second packet, half of the packets can also
be dropped by dropping N-packets followed by accepting N-packets.
Such a N-drop-N-accept pattern, resembles what RX-stages can acheive
by handling the XDP stage before netstack stage.

The XDP program must not be too small, which is currently is, because
what we are measuring there is the effect of keeping the XDP/eBPF-prog
hot in the I-cache, and also the early part of the driver code in I-cache.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>