samples/bpf: correctly accesss port_blacklist config and stats maps
Each entry in port_blacklist contains a bitfield of supported protocols
based on this enum:

enum {
	DDOS_FILTER_TCP = 0,
	DDOS_FILTER_UDP,
	DDOS_FILTER_MAX
};

An error in the previous implementation set fproto's value correctly for
the lookup in the BPF map, but did not have fproto set correctly for
selecting the stats map to update.  Without this fix stats are wrong.

Signed-off-by: Andy Gospodarek <gospo@broadcom.com>