xdp: fix xdp_bench01 action XDP_TX, such packets reach the TX wire
Simply bouncing the packet via XDP_TX is getting blocked by the
NIC hardware, in this case the ConnectX-4 driver mlx5.

The MAC-addresses need to be swapped, before the NIC HW allows
these packets to reach the wire.

For now, allow keeping the "invalid" case to allow further testing.
But print a big WARNING that packets might not reach the wire.

The correct use is enabled via using option --readmem and action XDP_TX.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>