docs: add area for Blogposts, Reports and Write-ups
This documentation area is used for publishing reports and write-ups
of my work, which I find relevant for other people.  This will closely
resemble previous my blogposts.  Hoping using this rst-text format
will make it less of an hassle to publish my work.

This directory is not intented to be integrated with the kernels
documentation tree.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>