samples/bpf: order of prog_fd[] must correspond with ELF order
It is highly problematic that the order of array prog_fd[] depend on
whether or not a bpf program section if referencing any maps or not.

I was implementing an XDP prog using tail calls, and it cause strange
issues that I could not count on the prog_fd[] order of the program,
when I was adjusting the programs...
This caused several days of wasted debugging! :-(

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>