samples/bpf: tc_bench01_redirect fix TC cmd usage of replace
The 'tc filter replace' command was creating a new filter per
invocation of the program, because a 'handle' and 'prio' was
not specified.

Thanks to Daniel Borkmann for spotting this bug.

Simply fix by specifying 'prio 1 handle 1'. 

Another fix would be to del + add on the clsact qdisc, as it 
would flush any filters attached.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>