samples/bpf: copy kernel library tools/lib/bpf/libbpf.{c,h}
Keep a git-repo copy of the kernel library for "bpf_object" ELF loading of eBPF files.

It is the hope that this lib can replace the bpf_load.c code.

If this lib becomes a success, it should instead be installed for the
kerne source tree, or be available via the Linux distros.  In that case, this
copy should be removed.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>