samples/bpf: building libbpf.a in Makefile
Add functionality to Makefile for building libbpf.a, but
this creates a dependency for the kernel source, as there is
a lot of include files from tools/include/ that is needed
to compile tools/lib/bpf/libbpf.c.

I'm not sure this is a good idea.  Notice, I've not added a
dependency that cause this to be build yet.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>