samples/bpf: xdp_redirect_cpu catch misusage like redirect to same CPU
The purpose of the cpumap is redirecting to another CPU than the RX CPU.
The use of the cpumap does require users to silo their IRQs and
kthread processing CPUs.

Redirecting to the same CPU defeats the purpose of the cpumap.  Thus, catch
such issues and expose it as a visible counter. Hopefully drawing attention
to the misconfig.

This commit also remove the "period" column.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>