samples/bpf: xdp_redirect_cpu add parameter for adding more CPUs --cpu
Just specify the --cpu option several times on the command line
to add more CPUs to the Round Robin CPU set.

Remeber that ixgbe have a default rx ring queue of 512, which is too
small and result in the page recycling failing.  This leads to bad
performance as the page allocator gets congested, via a spin_lock
in free_one_page().

Change the size via cmdline:
 ethtool -G ixgbe1 rx 1024 tx 1024

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>