samples/bpf: xdp_ddos01_blacklist handle parsing of double VLAN tagged packets
I'm controlled via Twitter:
 [1] https://twitter.com/majek04/status/1017492418227331073

People noticed that the code match on IEEE 802.1ad (ETH_P_8021AD) ethertype,
and this implies Q-in-Q or double tagged VLANs.  Thus, we better parse
the next VLAN header too.  It is even marked as a TODO.

This is related to CloudFlare blogpost:
 [2] https://blog.cloudflare.com/how-to-drop-10-million-packets/

I'm not doing to fix Marek's code as results are used blog:
 [3] https://github.com/cloudflare/cloudflare-blog/blob/master/2018-07-dropping-packets/xdp-drop-ebpf.c#L33-L41

Link: https://twitter.com/majek04/status/1017492418227331073
Reported-by: Florian Maury <florian.maury-cv@x-cli.eu>
Reported-by: Marek Majkowski <marek@cloudflare.com>
Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>