samples/bpf: xdp_vlan01 fix test script on veth driver
A common gotcha with XDP is that Linux can keep the VLAN tag
as a member in the SKB structure.  This is commonly known
as VLAN offloading.  This is because some NIC HW/drivers support
storing the VLAN id in the descriptor (both ingress and egress).

For the software driver veth, I forgot to disable TX VLAN offloading.

I usually always remember to disable RX VLAN offloading,
but because veth have two paired devices, there is also
a TX side that can do VLAN offloading.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>