samples/bpf: xdp_vlan01 workaround kernel bug in generic-XDP
here is a kernel bug for generic-XDP, that does not allow us to
remove a VLAN header, because skb->protocol still contain VLAN
ETH_P_8021Q indication, and this cause overwriting of our changes
via xdp_vlan_remove_outer2.

Thus, we cannot use use "xdp_vlan_remove_outerX" until this is
fixed in the kernel for generic-XDP.  This does work for native-XDP.

The workaround is to use "xdp_vlan_change" to change the VLAN id to zero,
as this have the same effect as removing the VLAN, seen from the kernel
point of view.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>