xdp_ddos01_blacklist_cmdline: Fix wrong usage of bpf_map_get_next_key
If inserting key 0 into hash, this code will cause issues listing IPs.

First change the variable name to be prev_key and key like:
  bpf_map_get_next_key(fd, prev_key, &key)

The prev_key need to be a real pointer, as we need to start with
prev_key = NULL.  Before prev_key was (a value on stack) implicit
zero. Thus, we basically asked to start the listing with key 0.
Thus, it is why it *looked* like it was all working when the hash
doesn't contain key=0, as then it fails-to-start-at-0 and kernel
provides the first-key.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>