bench_page_pool_cross_cpu: tweaking with small delays
It is primarily the concurrent return-path that I want to measure.
To get a proper result we want to see the case where return-path
is not stalling on an empty queue, resulting in spinning on the
ptr_ring_consume(queue) call.

Play with small nanosec delay's to get higher chance of
the queue isn't empty.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>